requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
grpcio==1.75.0
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.9
lazy_loader==0.4
lpips==0.1.4
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
pydantic==2.12.2
pydantic_core==2.41.4
pyparsing==3.2.4
pysodmetrics==1.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.16.2
segmentation_models_pytorch==0.5.0
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.2.1
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.9.20
timm==1.0.20
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
